+++ benchmark/utils.js
@@ -7,4 +7,4 @@ module.exports = {
 	getDataFile(filename, encoding = "utf8") {
 		return fs.readFileSync(path.join(__dirname, "data", filename), encoding);
 	}
-}
+};