+++ typings/index.d.ts
@@ -1,3 +1,0 @@
-/// <reference path="../index.d.ts" />
-/// <reference path="globals/jest/index.d.ts" />
-/// <reference path="globals/socket.io/index.d.ts" />